refactor: API 관련 코드(타입, 도메인 객체 변환 함수 등) cart.ts 와 product.ts 로 응집
